pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "jose-bot"
version = "0.0.1"
description = "Jose bot (Join confirm bot)"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "matrix-nio>=0.10.0",
    "PyYAML>=5.1.2",
    "xxhash",
]
classifiers=[
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[project.optional-dependencies]
dev = [
    "isort==5.0.4",
    "flake8==3.8.3",
    "flake8-comprehensions==3.2.3",
    "black==22.6.0",
]

[project.scripts]
jose-bot = "jose_bot:run"

# Setuptools
[tool.setuptools]
packages = ["jose_bot"]
